pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["boytsovau <alexbonch@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

classifiers = [
    "Topic :: Internet :: WWW/HTTP"
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"
requests = "^2.31.0"
bs4 = "^0.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]

app = "page_analyzer.app:index"
